IDL: Update discriminants to u32 instead of u8 (#54)
* IDL: Update discriminants to u32 instead of u8

#### Problem

The IDL used to generate the stake program client declares the
instruction discriminant as a `u8`, when the program actually uses
`u32`s for the discriminant.

#### Summary of changes

Update the IDL, update the generate script, then regenerate all of the
clients.

NOTE: This doesn't cover the `StakeState` and `StakeStateV2` enums,
which both use a `u32` discriminant too. I wasn't sure how to do that in
the IDL, so omitted it.

* Regenerate clients

* Update codama, remove trait overrides

* Update hooked file

* Regenerate client and lockfile

* Remove Eq trait, add u32 for stake state enum

* Regenerate clients